Extract KDocStyle class out of EndOfSentenceFormat.kt
